Scraper included into background service

+++ soundmoney/Models/DeepFinancial.cs
@@ -5,35 +5,35 @@
         public string Symbol { get; set; }
 
         // Header Metrics
-        public double CurrentPrice { get; set; }
-        public double MarketCapCr { get; set; }
-        public double BookValuePerShare { get; set; }
-        public double TotalSharesCr { get; set; }
-        public double ReportedRoePercent { get; set; }
-        public double DividendYieldPercent { get; set; }
+        public decimal CurrentPrice { get; set; }
+        public decimal MarketCapCr { get; set; }
+        public decimal BookValuePerShare { get; set; }
+        public decimal TotalSharesCr { get; set; }
+        public decimal ReportedRoePercent { get; set; }
+        public decimal DividendYieldPercent { get; set; }
 
         // P&L Metrics (Latest Year / TTM in Crores)
-        public double RevenueCr { get; set; }
-        public double OperatingProfitEbitdaCr { get; set; }
-        public double DepreciationCr { get; set; }
-        public double NetProfitCr { get; set; }
-        public double DividendPayoutPercent { get; set; }
+        public decimal RevenueCr { get; set; }
+        public decimal OperatingProfitEbitdaCr { get; set; }
+        public decimal DepreciationCr { get; set; }
+        public decimal NetProfitCr { get; set; }
+        public decimal DividendPayoutPercent { get; set; }
 
         // Balance Sheet Metrics (Latest Year in Crores)
-        public double ShareCapitalCr { get; set; }
-        public double ReservesCr { get; set; }
-        public double TotalEquityCr { get; set; }
-        public double TotalBorrowingsCr { get; set; }
-        public double NetFixedAssetsCr { get; set; }
-        public double CwipCr { get; set; }
-        public double CashAndEquivalentsCr { get; set; }
-        public double IntangibleAssetsCr { get; set; }
-        public double TotalLiabilitiesCr { get; set; }
-        public double TotalAssetsCr { get; set; }
+        public decimal ShareCapitalCr { get; set; }
+        public decimal ReservesCr { get; set; }
+        public decimal TotalEquityCr { get; set; }
+        public decimal TotalBorrowingsCr { get; set; }
+        public decimal NetFixedAssetsCr { get; set; }
+        public decimal CwipCr { get; set; }
+        public decimal CashAndEquivalentsCr { get; set; }
+        public decimal IntangibleAssetsCr { get; set; }
+        public decimal TotalLiabilitiesCr { get; set; }
+        public decimal TotalAssetsCr { get; set; }
 
         // Cash Flow Metrics (Latest Year in Crores)
-        public double CashFromOperationsCr { get; set; }
-        public double GrossCapexCr { get; set; }
-        public double FreeCashFlowCr { get; set; }
+        public decimal CashFromOperationsCr { get; set; }
+        public decimal GrossCapexCr { get; set; }
+        public decimal FreeCashFlowCr { get; set; }
     }
 }
